Added delete all user functionality + Segregate the controllers into User, Expense, Budget
