+++ src/main/java/com/tracker/expense_tracker/Service/UserService.java
@@ -11,4 +11,5 @@ public interface UserService {
     public String signup(SignUpDao sign);
     public List<User> getUser();
     public String login(LoginDao login);
+    public String deleteAll();
 }
